+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/view/model/GalleryListViewModel.kt
@@ -13,6 +13,8 @@ interface GalleryListViewModel {
     val itemScale: BehaviorSubject<GalleryItemScale>
     val selectedItemsCount: BehaviorSubject<Int>
     val selectedMediaByUid: LinkedHashMap<String, GalleryMedia>
+    val canAddSelectedToFavorites: Boolean
+    val canRemoveSelectedFromFavorites: Boolean
 
     fun postGalleryItemsAsync(repository: SimpleGalleryMediaRepository)
     fun postGalleryItems(repository: SimpleGalleryMediaRepository)